requirements.txt
appdirs==1.4.4
attrs==21.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
binwalk==2.3.3
blinker==1.7.0
Brotli==1.0.9
bs4==0.0.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
et-xmlfile==1.1.0
fake-useragent==1.5.0
filelock==3.16.1
Flask==3.0.2
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
html5lib==1.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
iniconfig==1.1.1
itsdangerous==2.1.2
jdcal==1.0
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.1.5
matplotlib==3.5.1
more-itertools==8.10.0
mpmath==0.0.0
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
packaging==21.3
pandas==2.2.3
Pillow==9.0.1
platformdirs==4.3.6
pluggy==0.13.0
ply==3.11
py==1.10.0
pycurl==7.45.3
pygame-ce==2.5.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pytest==6.2.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
pythran==0.10.0
PyTMX==3.32
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
sslib==0.2.0
sympy==1.9
systemd-python==234
toml==0.10.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.2.1
virtualenv==20.27.1
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.1
wfuzz==3.1.0
zipp==1.0.0
